created PaymentMethod component
