+++ src/pages/Payment/Payment.js
@@ -4,24 +4,18 @@ import { Redirect } from "react-router-dom";
 import clsx from "clsx";
 import { useFormik } from "formik";
 
-import RadioGroup from "@material-ui/core/RadioGroup";
 import FormControlLabel from "@material-ui/core/FormControlLabel";
 import Checkbox from "@material-ui/core/Checkbox";
 import Button from "../../components/Button";
 import UiInput from "../../components/UiInput";
 import UiInputDate from "../../components/UiInputDate";
-import UiCustomRadio from "../../components/UiCustomRadio";
+import PaymentMethod from "../../components/PaymentMethod";
 
 import paymentSchema from "./Payment-schema";
 
 import withCheckoutLayout from "../../hoc/withCheckoutLayout";
 import CheckoutContext from "../../context/checkout-context";
 
-import paypalImage from "../../img/icons/payment/Paypal.png";
-import applePayImage from "../../img/icons/payment/Apple.png";
-import visaImage from "../../img/icons/payment/Visa.png";
-import mastercardImage from "../../img/icons/payment/MasterCard.png";
-import amexImage from "../../img/icons/payment/AmericanExpress.png";
 import CVV from "../../img/icons/payment/CVV.svg";
 import CreditCard from "../../components/CreditCard";
 
@@ -49,27 +43,6 @@ function Payment() {
     },
   });
 
-  const paymentOptions = [
-    {
-      method: "card",
-      formText: "Credit/Debit Card",
-      formImage: false,
-      disabled: false,
-    },
-    {
-      method: "paypal",
-      formText: false,
-      formImage: paypalImage,
-      disabled: true,
-    },
-    {
-      method: "applePay",
-      formText: false,
-      formImage: applePayImage,
-      disabled: true,
-    },
-  ];
-
   function handleFlip() {
     return flip ? setFlip(false) : setFlip(true);
   }
@@ -80,46 +53,10 @@ function Payment() {
           <div className="row">
             <h3>Payment Details</h3>
             <hr />
-            <div className="row">
-              <h5>How would you like to pay?</h5>
-            </div>
-            <div className="row gy-2">
-              <RadioGroup
-                id="paymentMethod"
-                aria-label="payment method"
-                name="paymentMethod"
-                value={formik.values.paymentMethod}
-                onChange={formik.handleChange}
-                row
-              >
-                {paymentOptions.map((payment) => {
-                  return (
-                    <UiCustomRadio
-                      chosenValue={formik.values.paymentMethod}
-                      value={payment.method}
-                      formText={payment.formText}
-                      formImage={payment.formImage}
-                      disabled={payment.disabled}
-                      key={payment.method}
-                    />
-                  );
-                })}
-              </RadioGroup>
-            </div>
-            <div className="row gy-2">
-              <div className="col-12">
-                We accept the following debit/credit cards
-              </div>
-              <div className="col-12">
-                <img className="radio-border-box" src={visaImage} alt="" />
-                <img
-                  className="radio-border-box"
-                  src={mastercardImage}
-                  alt=""
-                />
-                <img className="radio-border-box" src={amexImage} alt="" />
-              </div>
-            </div>
+            <PaymentMethod
+              value={formik.values.paymentMethod}
+              changeHandler={formik.handleChange}
+            />
             <div className="row gy-4">
               <div className="col-6">
                 <UiInput
